refactor: extract custom useTheme hook to decouple global theme management ELUSoC_2026
